added lost rxn names + SBML and TXT model versions
Fixed a bug that removed rxn full names. For this changed functions:
* Matlab_Module/change_model/addArmReaction.m
* Matlab_Module/change_model/addEnzymesToRxn.m
* Matlab_Module/change_model/convertToEnzymeModel.m
* Matlab_Module/change_model/manualModifications.m

Also modified main script for saving model as .mat, .sbml and .txt. Both
ecYeast7 with no usage constrained and ecYeast7 with an overall enzyme
constraint fitted to batch data are available. For this:
* Changed function Matlab_Module/enhanceGEM.m
* Created function Models/saveECmodelSBML.m

All 6 versions of the model are available in the new folder 'Models'