fix: Added grRules and subsystems to rxns matched with enzymes
The grRules of the original reactions are matched to single enzymatic reactions and to arm reactions for isoenzymes cases.

Subsystems are copied from original reactions to every newly created one.

The rxnGeneMat is kept for the ecModel and it is updated with all the modifications that were done on the newRxns-genes relationships