feat: display RQ prediction error instead of excange fluxes
Some exchange fluxes are used as constraints (UB for GUR and LB for CO2 production), therefore it is not appropriate to use a median error in their prediction as a metric for evaluation of the ecModel performance.

Respiratory quotient is suggested instead due to the fact that oxygen is not being used  as constraint and the one imposed on CO2 production is not a hard one.